isNumber is fixed

+++ include/ac_common.php
@@ -78,7 +78,7 @@ function cmpDate ($p_date,$p_date_oth) {
 function isNumber($p_int) {
   if ( strlen (trim($p_int)) == 0 ) return 0;
 
-  if ( (string) $p_int === (string) (int) $p_int ) 
+  if ( (string) $p_int === (string) (float) $p_int ) 
     return 1;
   else
     return 0;
